Refactor the way we get CPU data.
We changed the way we get CPU data to account for inconsistent entries (less stats for a given timestamp etc.). We now ignore invalid values, resize stats for each core and return (as we did before) a single value for all cores.

When the user asks for a single value, we need to get two values from the database in order to calculate CPU utilization (according to the formula total_diff - idle_diff/total_diff)

minor change: we allow to user to specify step < 1000 but we just treat these as seconds (not millies)

TODO: we have to account for more than 16MB bson objects
as we use the aggregate command, we have to find a way to either deny queries that big, or find a way around this kind of queries