Added lower level disk stats (block).
We added disk stats to plot iops, octets and so on.

Fixed a couple of bugs (step == 0 and multiple fields in expression):

when we get step == 0 we replace it with the default step (60000).
when we get multiple requests (expression=cpu,memory,disk etc...) we
have to split the expression using split(',') because the [expression] hack
uses the whole string as a list item (e.g. cpu,memory = ['cpu,memory'])