Trasform nulls
An optional argument has been added to get_series method that
defines how null values returned by graphite are treated.

The default behaviour is to strip null values, but they can also
be left as is or replaced by some other defined value.

This is only used in get_stats where we transform null values
to 0. This was done as a (hopefully temporary) measure to fix
issues in d3 graphs in the mist.io frontend.