Use nonNegativeDerivative in Disk and Net targets
When we use derivative on a series of data that are actually a
counter and the counter wraps around its max value back to zero
or there are missing values, the derivative function makes long
negative spikes since the counter drops from a high value down
to zero (null's when values are missing are also treated as zeroes
when calculating derivatives). By replacing the derivative function
with nonNegativeDerivative we avoid negative values.